server/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Mobile Orienteering Server"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "google-auth>=2.37.0",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.0",
    "setuptools>=80.10.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "wheel>=0.46.2",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
server = "app.main:main"
